+++ x/domain/types/msgs.go
@@ -383,7 +383,7 @@ type MsgRenewAccount struct {
 	Domain string `json:"domain"`
 	// Name is the name of the account
 	Name string `json:"name"`
-	// Configurer is the signer of the request
+	// Signer is the signer of the request
 	Signer       sdk.AccAddress `json:"signer"`
 	FeePayerAddr sdk.AccAddress `json:"fee_payer"`
 }
@@ -435,7 +435,7 @@ func (m *MsgRenewAccount) GetSigners() []sdk.AccAddress {
 type MsgRenewDomain struct {
 	// Domain is the domain name to renew
 	Domain string `json:"domain"`
-	// Configurer is the request signer
+	// Signer is the request signer
 	Signer       sdk.AccAddress `json:"signer"`
 	FeePayerAddr sdk.AccAddress `json:"fee_payer"`
 }