+++ x/configuration/types/fees_test.go
@@ -1,24 +1,24 @@
 package types
 
 import (
-	"github.com/cosmos/cosmos-sdk/types"
 	"testing"
+
+	"github.com/cosmos/cosmos-sdk/types"
 )
 
 func TestFees_Validate(t *testing.T) {
 	type fields struct {
 		FeeCoinDenom                 string
 		FeeCoinPrice                 types.Dec
-		DefaultFee                   types.Dec
-		RegisterClosedAccount        types.Dec
-		RegisterOpenAccount          types.Dec
-		TransferClosedAccount        types.Dec
-		TransferOpenAccount          types.Dec
+		FeeDefault                   types.Dec
+		RegisterAccountClosed        types.Dec
+		RegisterAccountOpen          types.Dec
+		TransferAccountClosed        types.Dec
+		TransferAccountOpen          types.Dec
 		ReplaceAccountTargets        types.Dec
 		AddAccountCertificate        types.Dec
 		DelAccountCertificate        types.Dec
 		SetAccountMetadata           types.Dec
-		RegisterDomain               types.Dec
 		RegisterDomain1              types.Dec
 		RegisterDomain2              types.Dec
 		RegisterDomain3              types.Dec
@@ -28,7 +28,7 @@ func TestFees_Validate(t *testing.T) {
 		RegisterOpenDomainMultiplier types.Dec
 		TransferDomainClosed         types.Dec
 		TransferDomainOpen           types.Dec
-		RenewOpenDomain              types.Dec
+		RenewDomainOpen              types.Dec
 	}
 	tests := []struct {
 		name    string
@@ -55,16 +55,15 @@ func TestFees_Validate(t *testing.T) {
 			f := &Fees{
 				FeeCoinDenom:                 tt.fields.FeeCoinDenom,
 				FeeCoinPrice:                 tt.fields.FeeCoinPrice,
-				DefaultFee:                   tt.fields.DefaultFee,
-				RegisterClosedAccount:        tt.fields.RegisterClosedAccount,
-				RegisterOpenAccount:          tt.fields.RegisterOpenAccount,
-				TransferClosedAccount:        tt.fields.TransferClosedAccount,
-				TransferOpenAccount:          tt.fields.TransferOpenAccount,
+				FeeDefault:                   tt.fields.FeeDefault,
+				RegisterAccountClosed:        tt.fields.RegisterAccountClosed,
+				RegisterAccountOpen:          tt.fields.RegisterAccountOpen,
+				TransferAccountClosed:        tt.fields.TransferAccountClosed,
+				TransferAccountOpen:          tt.fields.TransferAccountOpen,
 				ReplaceAccountTargets:        tt.fields.ReplaceAccountTargets,
 				AddAccountCertificate:        tt.fields.AddAccountCertificate,
 				DelAccountCertificate:        tt.fields.DelAccountCertificate,
 				SetAccountMetadata:           tt.fields.SetAccountMetadata,
-				RegisterDomain:               tt.fields.RegisterDomain,
 				RegisterDomain1:              tt.fields.RegisterDomain1,
 				RegisterDomain2:              tt.fields.RegisterDomain2,
 				RegisterDomain3:              tt.fields.RegisterDomain3,
@@ -74,7 +73,7 @@ func TestFees_Validate(t *testing.T) {
 				RegisterOpenDomainMultiplier: tt.fields.RegisterOpenDomainMultiplier,
 				TransferDomainClosed:         tt.fields.TransferDomainClosed,
 				TransferDomainOpen:           tt.fields.TransferDomainOpen,
-				RenewOpenDomain:              tt.fields.RenewOpenDomain,
+				RenewDomainOpen:              tt.fields.RenewDomainOpen,
 			}
 			if err := f.Validate(); (err != nil) != tt.wantErr {
 				t.Errorf("Validate() error = %v, wantErr %v", err, tt.wantErr)